🎨 refactor: shared 컴포넌트 모듈 model / ui 분리
