+++ src/shared/Button/model/types.ts
@@ -1,0 +1,31 @@
+import { cva, VariantProps } from "class-variance-authority"
+
+export const buttonVariants = cva(
+    "inline-flex items-center justify-center rounded-md text-sm font-medium transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:opacity-50 disabled:pointer-events-none ring-offset-background",
+    {
+        variants: {
+            variant: {
+            default: "bg-blue-500 text-white hover:bg-blue-600",
+            destructive: "bg-red-500 text-white hover:bg-red-600",
+            outline: "border border-gray-300 bg-transparent text-gray-700 hover:bg-gray-100",
+            secondary: "bg-gray-200 text-gray-800 hover:bg-gray-300",
+            ghost: "bg-transparent text-gray-700 hover:bg-gray-100",
+            link: "underline-offset-4 hover:underline text-blue-500",
+            },
+            size: {
+            default: "h-10 py-2 px-4",
+            sm: "h-8 px-3 rounded-md text-xs",
+            lg: "h-11 px-8 rounded-md",
+            icon: "h-9 w-9",
+            },
+        },
+        defaultVariants: {
+            variant: "default",
+            size: "default",
+        },
+    },
+)
+
+export interface ButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement>, VariantProps<typeof buttonVariants> {
+    className?: string
+}
