+++ src/shared/Button/ui/Button.tsx
@@ -1,0 +1,8 @@
+import { forwardRef } from "react"
+import { buttonVariants, ButtonProps } from "../model/types"
+
+export const Button = forwardRef<HTMLButtonElement, ButtonProps>(({ className, variant, size, ...props }, ref) => {
+    return <button className={buttonVariants({ variant, size, className })} ref={ref} {...props} />
+})
+
+Button.displayName = "Button"
