+++ src/shared/Card/model/types.ts
@@ -1,0 +1,4 @@
+export interface CardProps {
+  className?: string
+  children?: React.ReactNode | string
+}
