+++ src/shared/Card/ui/Card.tsx
@@ -1,0 +1,30 @@
+import { forwardRef } from "react"
+import { CardProps } from "../model/types.ts"
+
+// 카드 컴포넌트
+const Card = forwardRef<HTMLInputElement, CardProps>(({ className, ...props }, ref) => (
+  <div ref={ref} className={`rounded-lg border bg-card text-card-foreground shadow-sm ${className}`} {...props} />
+))
+Card.displayName = "Card"
+
+const CardHeader = forwardRef<HTMLInputElement, CardProps>(({ className, ...props }, ref) => (
+  <div ref={ref} className={`flex flex-col space-y-1.5 p-6 ${className}`} {...props} />
+))
+CardHeader.displayName = "CardHeader"
+
+const CardTitle = forwardRef<HTMLInputElement, CardProps>(({ className, ...props }, ref) => (
+  <h3 ref={ref} className={`text-2xl font-semibold leading-none tracking-tight ${className}`} {...props} />
+))
+CardTitle.displayName = "CardTitle"
+
+const CardContent = forwardRef<HTMLInputElement, CardProps>(({ className, ...props }, ref) => (
+  <div ref={ref} className={`p-6 pt-0 ${className}`} {...props} />
+))
+CardContent.displayName = "CardContent"
+
+export {
+  Card,
+  CardHeader,
+  CardTitle,
+  CardContent,
+}
