+++ src/shared/Dialog/model/types.ts
@@ -1,0 +1,4 @@
+export interface DialogProps {
+  className?: string
+  children?: React.ReactNode | string
+}
