+++ src/shared/Select/model/types.ts
@@ -1,0 +1,6 @@
+export interface SelectProps {
+  className?: string
+  children?: React.ReactNode | string
+  position?: "popper" | "item-aligned"
+  value?: string
+};
