+++ src/shared/Table/model/types.ts
@@ -1,0 +1,5 @@
+// 테이블 컴포넌트
+export interface TableProps {
+  className?: string
+  children?: React.ReactNode | string
+}
