+++ src/shared/textarea/model/types.ts
@@ -1,0 +1,8 @@
+// 텍스트 영역 컴포넌트
+export interface TextareaProps {
+  rows?: number;
+  className?: string;
+  placeholder?: string;
+  value?: string;
+  onChange?: (e: React.ChangeEvent<HTMLTextAreaElement>) => void;  // onChange 명확화
+}
